moved the network models into their own files under a models directory
