+++ Sources/KlaviyoCore/InternalAPIModels.swift
@@ -10,358 +10,6 @@ import AnyCodable
 import Foundation
 
 public enum KlaviyoEndpoint: Equatable, Codable {
-    public struct CreateProfilePayload: Equatable, Codable {
-        public init(data: KlaviyoEndpoint.CreateProfilePayload.Profile) {
-            self.data = data
-        }
-
-        /**
-         Internal structure which has details not needed by the API.
-         */
-        public struct Profile: Equatable, Codable {
-            var type = "profile"
-            public struct Attributes: Equatable, Codable {
-                public let email: String?
-                public let phoneNumber: String?
-                public let externalId: String?
-                public let anonymousId: String
-                public var firstName: String?
-                public var lastName: String?
-                public var organization: String?
-                public var title: String?
-                public var image: String?
-                public var location: PublicProfile.Location?
-                public var properties: AnyCodable
-                enum CodingKeys: String, CodingKey {
-                    case email
-                    case phoneNumber = "phone_number"
-                    case externalId = "external_id"
-                    case anonymousId = "anonymous_id"
-                    case firstName = "first_name"
-                    case lastName = "last_name"
-                    case organization
-                    case title
-                    case image
-                    case location
-                    case properties
-                }
-
-                public init(attributes: PublicProfile,
-                            anonymousId: String) {
-                    email = attributes.email
-                    phoneNumber = attributes.phoneNumber
-                    externalId = attributes.externalId
-                    firstName = attributes.firstName
-                    lastName = attributes.lastName
-                    organization = attributes.organization
-                    title = attributes.title
-                    image = attributes.image
-                    location = attributes.location
-                    properties = AnyCodable(attributes.properties)
-                    self.anonymousId = anonymousId
-                }
-            }
-
-            public var attributes: Attributes
-            public init(profile: PublicProfile, anonymousId: String) {
-                attributes = Attributes(
-                    attributes: profile,
-                    anonymousId: anonymousId)
-            }
-
-            public init(attributes: Attributes) {
-                self.attributes = attributes
-            }
-        }
-
-        public var data: Profile
-    }
-
-    public struct CreateEventPayload: Equatable, Codable {
-        public struct Event: Equatable, Codable {
-            public struct Attributes: Equatable, Codable {
-                public struct Metric: Equatable, Codable {
-                    public let data: MetricData
-
-                    public struct MetricData: Equatable, Codable {
-                        var type: String = "metric"
-
-                        public let attributes: MetricAttributes
-
-                        public init(name: String) {
-                            attributes = .init(name: name)
-                        }
-
-                        public struct MetricAttributes: Equatable, Codable {
-                            public let name: String
-                        }
-                    }
-
-                    public init(name: String) {
-                        data = .init(name: name)
-                    }
-                }
-
-                public struct Profile: Equatable, Codable {
-                    public let data: CreateProfilePayload.Profile
-
-                    public init(attributes: PublicProfile,
-                                anonymousId: String) {
-                        data = .init(profile: attributes, anonymousId: anonymousId)
-                    }
-                }
-
-                public let metric: Metric
-                public var properties: AnyCodable
-                public let profile: Profile
-                public let time: Date
-                public let value: Double?
-                public let uniqueId: String
-                public init(attributes: PublicEvent,
-                            anonymousId: String? = nil) {
-                    metric = Metric(name: attributes.metric.name.value)
-                    properties = AnyCodable(attributes.properties)
-                    value = attributes.value
-                    time = attributes.time
-                    uniqueId = attributes.uniqueId
-
-                    profile = .init(attributes: .init(
-                        email: attributes.identifiers?.email,
-                        phoneNumber: attributes.identifiers?.phoneNumber,
-                        externalId: attributes.identifiers?.externalId),
-                    anonymousId: anonymousId ?? "")
-                }
-
-                enum CodingKeys: String, CodingKey {
-                    case metric
-                    case properties
-                    case profile
-                    case time
-                    case value
-                    case uniqueId = "unique_id"
-                }
-            }
-
-            var type = "event"
-            public var attributes: Attributes
-            public init(event: PublicEvent,
-                        anonymousId: String? = nil) {
-                attributes = .init(attributes: event, anonymousId: anonymousId)
-            }
-        }
-
-        mutating func appendMetadataToProperties() {
-            let context = KlaviyoAPI._appContextInfo
-            // TODO: Fixme
-//                let metadata: [String: Any] = [
-//                    "Device ID": context.deviceId,
-//                    "Device Manufacturer": context.manufacturer,
-//                    "Device Model": context.deviceModel,
-//                    "OS Name": context.osName,
-//                    "OS Version": context.osVersion,
-//                    "SDK Name": __klaviyoSwiftName,
-//                    "SDK Version": __klaviyoSwiftVersion,
-//                    "App Name": context.appName,
-//                    "App ID": context.bundleId,
-//                    "App Version": context.appVersion,
-//                    "App Build": context.appBuild,
-//                    "Push Token": analytics.state().pushTokenData?.pushToken as Any
-//                ]
-
-            let metadata = [String: Any]()
-            let originalProperties = data.attributes.properties.value as? [String: Any] ?? [:]
-            data.attributes.properties = AnyCodable(originalProperties.merging(metadata) { _, new in new })
-        }
-
-        public var data: Event
-        public init(data: Event) {
-            self.data = data
-        }
-    }
-
-    public struct PushTokenPayload: Equatable, Codable {
-        public init(data: KlaviyoEndpoint.PushTokenPayload.PushToken) {
-            self.data = data
-        }
-
-        public let data: PushToken
-
-        public init(pushToken: String,
-                    enablement: String,
-                    background: String,
-                    profile: PublicProfile,
-                    anonymousId: String) {
-            data = .init(
-                pushToken: pushToken,
-                enablement: enablement,
-                background: background,
-                profile: profile,
-                anonymousId: anonymousId)
-        }
-
-        public struct PushToken: Equatable, Codable {
-            var type = "push-token"
-            public var attributes: Attributes
-
-            public init(pushToken: String,
-                        enablement: String,
-                        background: String,
-                        profile: PublicProfile,
-                        anonymousId: String) {
-                attributes = .init(
-                    pushToken: pushToken,
-                    enablement: enablement,
-                    background: background,
-                    profile: profile,
-                    anonymousId: anonymousId)
-            }
-
-            public struct Attributes: Equatable, Codable {
-                public let profile: Profile
-                public let token: String
-                public let enablementStatus: String
-                public let backgroundStatus: String
-                public let deviceMetadata: MetaData
-                public let platform: String = "ios"
-                public let vendor: String = "APNs"
-
-                enum CodingKeys: String, CodingKey {
-                    case token
-                    case platform
-                    case enablementStatus = "enablement_status"
-                    case profile
-                    case vendor
-                    case backgroundStatus = "background"
-                    case deviceMetadata = "device_metadata"
-                }
-
-                public init(pushToken: String,
-                            enablement: String,
-                            background: String,
-                            profile: PublicProfile,
-                            anonymousId: String) {
-                    token = pushToken
-
-                    enablementStatus = enablement
-                    backgroundStatus = background
-                    self.profile = .init(attributes: profile, anonymousId: anonymousId)
-                    deviceMetadata = .init(context: KlaviyoAPI._appContextInfo)
-                }
-
-                public struct Profile: Equatable, Codable {
-                    public let data: CreateProfilePayload.Profile
-
-                    public init(attributes: PublicProfile,
-                                anonymousId: String) {
-                        data = .init(profile: attributes, anonymousId: anonymousId)
-                    }
-                }
-
-                public struct MetaData: Equatable, Codable {
-                    public let deviceId: String
-                    public let deviceModel: String
-                    public let manufacturer: String
-                    public let osName: String
-                    public let osVersion: String
-                    public let appId: String
-                    public let appName: String
-                    public let appVersion: String
-                    public let appBuild: String
-                    public let environment: String
-                    public let klaviyoSdk: String
-                    public let sdkVersion: String
-
-                    enum CodingKeys: String, CodingKey {
-                        case deviceId = "device_id"
-                        case klaviyoSdk = "klaviyo_sdk"
-                        case sdkVersion = "sdk_version"
-                        case deviceModel = "device_model"
-                        case osName = "os_name"
-                        case osVersion = "os_version"
-                        case manufacturer
-                        case appName = "app_name"
-                        case appVersion = "app_version"
-                        case appBuild = "app_build"
-                        case appId = "app_id"
-                        case environment
-                    }
-
-                    public init(context: AppContextInfo) {
-                        deviceId = context.deviceId
-                        deviceModel = context.deviceModel
-                        manufacturer = context.manufacturer
-                        osName = context.osName
-                        osVersion = context.osVersion
-                        appId = context.bundleId
-                        appName = context.appName
-                        appVersion = context.appVersion
-                        appBuild = context.appBuild
-                        environment = context.environment
-                        klaviyoSdk = __klaviyoSwiftName
-                        sdkVersion = __klaviyoSwiftVersion
-                    }
-                }
-            }
-        }
-    }
-
-    public struct UnregisterPushTokenPayload: Equatable, Codable {
-        public let data: PushToken
-
-        public init(pushToken: String,
-                    profile: PublicProfile,
-                    anonymousId: String) {
-            data = .init(
-                pushToken: pushToken,
-                profile: profile,
-                anonymousId: anonymousId)
-        }
-
-        public struct PushToken: Equatable, Codable {
-            var type = "push-token-unregister"
-            public var attributes: Attributes
-
-            public init(pushToken: String,
-                        profile: PublicProfile,
-                        anonymousId: String) {
-                attributes = .init(
-                    pushToken: pushToken,
-                    profile: profile,
-                    anonymousId: anonymousId)
-            }
-
-            public struct Attributes: Equatable, Codable {
-                public let profile: Profile
-                public let token: String
-                public let platform: String = "ios"
-                public let vendor: String = "APNs"
-
-                enum CodingKeys: String, CodingKey {
-                    case token
-                    case platform
-                    case profile
-                    case vendor
-                }
-
-                public init(pushToken: String,
-                            profile: PublicProfile,
-                            anonymousId: String) {
-                    token = pushToken
-                    self.profile = .init(attributes: profile, anonymousId: anonymousId)
-                }
-
-                public struct Profile: Equatable, Codable {
-                    public let data: CreateProfilePayload.Profile
-
-                    public init(attributes: PublicProfile,
-                                anonymousId: String) {
-                        data = .init(profile: attributes, anonymousId: anonymousId)
-                    }
-                }
-            }
-        }
-    }
-
     case createProfile(CreateProfilePayload)
     case createEvent(CreateEventPayload)
     case registerPushToken(PushTokenPayload)
